refactor: organize api types
